requirements.txt
click
pandas
pydantic
pydantic-settings
PyYAML
tqdm
openpyxl
XlsxWriter
Jinja2
pyecharts